pi-daemon/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
opencv-python-headless==4.10.0.84
evdev==1.7.1
Pillow==10.4.0
